+++ pdf/test/widget_table_test.dart
@@ -145,6 +145,47 @@ void main() {
     ));
   });
 
+  test('Table Widget TableCellVerticalAlignment', () {
+    pdf.addPage(
+      MultiPage(
+        build: (Context context) {
+          return <Widget>[
+            Table(
+              defaultColumnWidth: const FixedColumnWidth(20),
+              children: List<TableRow>.generate(
+                TableCellVerticalAlignment.values.length,
+                (int index) {
+                  final TableCellVerticalAlignment align =
+                      TableCellVerticalAlignment.values[
+                          index % TableCellVerticalAlignment.values.length];
+
+                  return TableRow(
+                    verticalAlignment: align,
+                    children: <Widget>[
+                      Container(
+                        child: Text('Vertical'),
+                        color: PdfColors.red,
+                      ),
+                      Container(
+                        child: Text('alignment $index'),
+                        color: PdfColors.yellow,
+                        height: 60,
+                      ),
+                      Container(
+                        child: Text(align.toString().substring(27)),
+                        color: PdfColors.green,
+                      ),
+                    ],
+                  );
+                },
+              ),
+            ),
+          ];
+        },
+      ),
+    );
+  });
+
   tearDownAll(() {
     final File file = File('widgets-table.pdf');
     file.writeAsBytesSync(pdf.save());